@@ -12,48 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.5",
-    "@angular/common": "^17.3.5",
-    "@angular/compiler": "^17.3.5",
-    "@angular/core": "^17.3.5",
-    "@angular/forms": "^17.3.5",
-    "@angular/localize": "^17.3.5",
-    "@angular/platform-browser": "^17.3.5",
-    "@angular/platform-browser-dynamic": "^17.3.5",
-    "@angular/router": "^17.3.5",
-    "@angular/service-worker": "^17.3.5",
+    "@angular/animations": "^19.2.1",
+    "@angular/common": "^19.2.1",
+    "@angular/compiler": "^19.2.1",
+    "@angular/core": "^19.2.1",
+    "@angular/forms": "^19.2.1",
+    "@angular/localize": "^19.2.1",
+    "@angular/platform-browser": "^19.2.1",
+    "@angular/platform-browser-dynamic": "^19.2.1",
+    "@angular/router": "^19.2.1",
+    "@angular/service-worker": "^19.2.1",
     "@digitalspace/bcparks-bootstrap-theme": "^1.4.6",
-    "@moddi3/ngx-intl-tel-input": "^3.3.7",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@moddi3/ngx-intl-tel-input": "^16.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^17",
     "@popperjs/core": "^2.11.8",
-    "angularx-qrcode": "^17.0.0",
+    "angularx-qrcode": "^18",
     "google-libphonenumber": "^3.2.34",
     "guid-typescript": "^1.0.9",
-    "intl-tel-input": "^18.2.1",
+    "intl-tel-input": "^19",
     "jwt-decode": "^4.0.0",
     "luxon": "^3.4.4",
     "material-icons": "^1.13.12",
-    "ngx-bootstrap": "12.0.0",
+    "ngx-bootstrap": "^18",
     "ngx-pagination": "^6.0.3",
     "ngx-toastr": "^18.0.0",
     "ngx-turnstile": "^0.2.1",
     "rxjs": "~7.8.1",
-    "zone.js": "~0.14.4"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.5",
-    "@angular-eslint/builder": "17.3.0",
-    "@angular-eslint/eslint-plugin": "17.3.0",
-    "@angular-eslint/eslint-plugin-template": "17.3.0",
-    "@angular-eslint/schematics": "17.3.0",
-    "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "^17.3.5",
-    "@angular/compiler-cli": "^17.3.5",
+    "@angular-devkit/build-angular": "^19.2.1",
+    "@angular-eslint/builder": "18.4.3",
+    "@angular-eslint/eslint-plugin": "18.4.3",
+    "@angular-eslint/eslint-plugin-template": "18.4.3",
+    "@angular-eslint/schematics": "18.4.3",
+    "@angular-eslint/template-parser": "18.4.3",
+    "@angular/cli": "^19.2.1",
+    "@angular/compiler-cli": "^19.2.1",
     "@types/jasmine": "^4.3.1",
     "@types/node": "^18.16.2",
-    "@typescript-eslint/eslint-plugin": "5.59.1",
-    "@typescript-eslint/parser": "5.59.1",
-    "eslint": "^8.28.0",
+    "@typescript-eslint/eslint-plugin": "7.11.0",
+    "@typescript-eslint/parser": "7.11.0",
+    "@typescript-eslint/types": "^8",
+    "@typescript-eslint/utils": "^8",
+    "eslint": "8.57.0",
     "http-server": "^14.1.1",
     "jasmine-core": "^4.6.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -65,6 +67,7 @@
     "karma-spec-reporter": "^0.0.36",
     "ts-node": "~10.9.1",
     "tslib": "^2.4.0",
-    "typescript": "5.4.5"
+    "typescript": "5.8.2",
+    "vite": "^5"
   }
 }
