@@ -7,9 +7,10 @@ import { ToastService } from './services/toast.service';
 import { Constants } from './shared/utils/constants';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    standalone: false
 })
 export class AppComponent implements OnInit {
   public breadcrumbs: IBreadcrumb[];
