@@ -2,9 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ConfigService } from '../shared/services/config.service';
 
 @Component({
-  selector: 'app-footer',
-  templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss']
+    selector: 'app-footer',
+    templateUrl: './footer.component.html',
+    styleUrls: ['./footer.component.scss'],
+    standalone: false
 })
 export class FooterComponent implements OnInit {
   public hashVersion: string;
