@@ -2,9 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ConfigService } from '../shared/services/config.service';
 
 @Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss']
+    selector: 'app-header',
+    templateUrl: './header.component.html',
+    styleUrls: ['./header.component.scss'],
+    standalone: false
 })
 export class HeaderComponent implements OnInit {
   public envName: string;
