@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-guidelines',
-  templateUrl: './guidelines.component.html',
-  styleUrls: ['./guidelines.component.scss']
+    selector: 'app-guidelines',
+    templateUrl: './guidelines.component.html',
+    styleUrls: ['./guidelines.component.scss'],
+    standalone: false
 })
 export class GuidelinesComponent {}
