@@ -2,9 +2,10 @@ import { Component, Input, OnInit } from '@angular/core';
 import { ConfigService } from 'src/app/shared/services/config.service';
 
 @Component({
-  selector: 'app-park-details',
-  templateUrl: './park-details.component.html',
-  styleUrls: ['./park-details.component.scss']
+    selector: 'app-park-details',
+    templateUrl: './park-details.component.html',
+    styleUrls: ['./park-details.component.scss'],
+    standalone: false
 })
 export class ParkDetailsComponent implements OnInit {
   @Input() park;
