@@ -3,9 +3,10 @@ import { Component, Input } from '@angular/core';
 import { AlertObject } from './alert-object';
 
 @Component({
-  selector: 'app-alert',
-  templateUrl: './alert.component.html',
-  styleUrls: ['./alert.component.scss']
+    selector: 'app-alert',
+    templateUrl: './alert.component.html',
+    styleUrls: ['./alert.component.scss'],
+    standalone: false
 })
 export class AlertComponent {
   @Input() data: AlertObject;
