@@ -16,9 +16,10 @@ import { takeUntil } from 'rxjs/operators';
 import { Utils } from '../../utils/utils';
 
 @Component({
-  selector: 'app-date-picker',
-  templateUrl: './date-picker.component.html',
-  styleUrls: ['./date-picker.component.scss']
+    selector: 'app-date-picker',
+    templateUrl: './date-picker.component.html',
+    styleUrls: ['./date-picker.component.scss'],
+    standalone: false
 })
 export class DatePickerComponent implements OnInit, OnChanges, OnDestroy {
   @Input() control: UntypedFormControl;
