@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-important-booking-info',
-  templateUrl: './important-booking-info.component.html',
-  styleUrls: ['./important-booking-info.component.scss']
+    selector: 'app-important-booking-info',
+    templateUrl: './important-booking-info.component.html',
+    styleUrls: ['./important-booking-info.component.scss'],
+    standalone: false
 })
 export class ImportantBookingInfoComponent {}
