@@ -3,9 +3,10 @@ import { TableObject } from '../table-template/table-object';
 import { ITableMessage } from '../table-template/table-row-component';
 
 @Component({
-  selector: 'app-list',
-  templateUrl: './list.component.html',
-  styleUrls: ['./list.component.scss']
+    selector: 'app-list',
+    templateUrl: './list.component.html',
+    styleUrls: ['./list.component.scss'],
+    standalone: false
 })
 export class ListComponent implements OnChanges {
   @Input() tableRowComponent;
