@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-tips',
-  templateUrl: './tips.component.html',
-  styleUrls: ['./tips.component.scss']
+    selector: 'app-tips',
+    templateUrl: './tips.component.html',
+    styleUrls: ['./tips.component.scss'],
+    standalone: false
 })
 export class TipsComponent {}
