@@ -11,8 +11,8 @@
   window.__env.ENVIRONMENT = 'local';
 
   window.__env.API_LOCATION = 'http://localhost:3000';
-  window.__env.API_PATH = '/';
-  window.__env.API_PUBLIC_PATH = '/';
+  window.__env.API_PATH = '/api';
+  window.__env.API_PUBLIC_PATH = '/api';
   window.__env.ASSETS_S3_URL = 'https://d3ptawfth8sh71.cloudfront.net';
 
 
